\name{mpsolve}
\alias{mpsolve}
\title{
  Solve a MetaPost Path
}
\description{
  Generates a set of cubic Bezier curve control points
  from a MetaPost path description.
}
\usage{
mpsolve(x)
}
\arguments{
  \item{x}{
    A MetaPost path, as generated by \code{\link{knot}} etc.
  }
}
\value{
  A list of Bezier control points (\code{mpcontrols} objects).
}
\author{
  Paul Murrell
}
\seealso{
  \code{\link{knot}},
  \code{\link{grid.mplib}}
}
\examples{
options(metapost.units="in")
p <- knot(0, 0) + dir(0) + dir(0) + knot(1, 1)
controls <- mpsolve(p)
controls
grid.mplib(p)
}
\keyword{ dplot }% use one of  RShowDoc("KEYWORDS")
